backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
boto3==1.35.0
torch>=2.1.0
transformers>=4.36.0
pytorch-crf>=0.7.2
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
nltk>=3.8.0
pydantic>=2.0.0
python-multipart>=0.0.9
requests>=2.31.0
pytest>=8.0.0
httpx>=0.27.0
